Move to use Spell Finder instead of filter_builder
Goal: simplify and clean up logic around spell finding
- First alphabetize the gemfile cause the rubocop was mad at me
- add in spring commands for rspec (running `spring rspec` now should
start a spring process and speed up boot time for specs)
- Add some default scopes to the spells model, we should always be
ordering by name alphabetized and have a limit of 50 (instead of getting
all the values and then paginating an array)
- Move spell filter_builder to a new model called SpellFinder which I
think is a more accurate name for what we want it to do. It's almost
like a repository model
- Get rid of inheritence in favor of using composable params
- Update the specs to use `match_array` which is a way of comparing
arrays when they should really be sets (Sets orders don't matter).
Although I should change this cause the order does matter and it should
be alphabetized
- Update the controller and search form to use more accurate names for
params

Ultimately I think our goal will be to build something that allows us to
use an actual filter string that will let us parse out and compose much
more complex queries accross possible types without the constant if
checking. We could use something like
https://kschiess.github.io/parslet/ to parse a string built out of all
the possible limits.

You could imagine something where we say "Give me all the level five
spells that belong to paladins or clerics that begin with the letter S
or begin with the letter F" which would be cool but MVP this shit dude

Also in rails 5 everything under the models dir is loaded by default so
if we want to have a service folder it should be `app/models/services`